+++ src/index.js
@@ -4,8 +4,8 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter as Router } from "react-router-dom";
-import {DataProvider} from "../src/contexts/datafetch-context"
-
+import { DataProvider } from "../src/contexts/datafetch-context";
+import { ProductProvider } from "./contexts/productListing-context";
 
 // Call make Server
 makeServer();
@@ -14,8 +14,10 @@ ReactDOM.render(
   <React.StrictMode>
     <Router>
       <DataProvider>
-        <App />
-      </DataProvider>  
+        <ProductProvider>
+          <App />
+        </ProductProvider>
+      </DataProvider>
     </Router>
   </React.StrictMode>,
   document.getElementById("root")
