+++ src/utils/initialFilters.js
@@ -1,0 +1,13 @@
+
+// this is where the filters would go back on clear
+
+const initialFilters = {
+    sortItemsBy: null,
+    rateBy: null,
+    inStock: false,
+    fastDelivery: false,
+    category: [], // we pass an empty array here so that multiple categories can be put in an array (MEN, Women, Kids, etc..) (for us to be able to select multiple categories)
+    priceRange: 1990
+  };
+
+  export {initialFilters}
