+++ src/utils/sortByPrice.js
@@ -1,6 +1,11 @@
 //function definition to sort items by price
 
-const getSortedItems = (productData, sortItemsBy) => {
+import { useProduct } from "../contexts/datafetch-context";
+
+
+
+  const getSortedItems = (productData, sortItemsBy) => {
+
     if (sortItemsBy === "LOW_TO_HIGH") {
       return [...productData].sort((item1, item2) => {
         const price1 = item1.price - item2.price;
@@ -8,11 +13,15 @@ const getSortedItems = (productData, sortItemsBy) => {
       });
     } else if (sortItemsBy === "HIGH_TO_LOW") {
       return [...productData].sort((item1, item2) => {
-        const price2 = item2.price - item1.price;
+        const price2 = item2.price - item1.price; //the price here is what we get from the db
         return price2;
       });
     }
     return productData;
   };
 
-  export {getSortedItems};
+  const {products} = useProduct();
+
+  const getSortedList = getSortedItems(products, state.sortItemsBy); // 1
+
+  export {getSortedItems, getSortedList}
